+++ xtimetracker/cli/__init__.py
@@ -2,12 +2,13 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from . import cli  # noqa
-from .add import add  # noqa
-from .cli import aggregate  # noqa
-from .cli import log  # noqa
-from .cli import report  # noqa
-from .constants import SHORTCUT_OPTIONS  # noqa
-from .start import start  # noqa
-from .status import status  # noqa
-from .stop import stop  # noqa
+from . import cli
+from .add import add
+from .cancel import cancel
+from .cli import aggregate
+from .cli import log
+from .cli import report
+from .constants import SHORTCUT_OPTIONS
+from .start import start
+from .status import status
+from .stop import stop