+++ xtimetracker/cli/stop.py
@@ -19,39 +19,23 @@ if TYPE_CHECKING:
     from ..timetracker import TimeTracker
 
 
-@cli.command(context_settings={"ignore_unknown_options": True})
-@click.option(
-    "-c",
-    "--cancel",
-    is_flag=True,
-    default=False,
-    help="Cancel current project monitoring.",
-)
+@cli.command()
 @click.pass_obj
 @catch_timetracker_error
-def stop(tt: TimeTracker, cancel):
+def stop(tt: TimeTracker):
     """
-    Stop or cancel monitoring time for the current project.
+    Stop tracking current activity.
     """
-    if cancel:
-        old = tt.cancel()
-        click.echo(
-            "Canceling current monitoring for project {}{}".format(
-                style("project", old["project"]),
-                (" " if old["tags"] else "") + style("tags", old["tags"]),
-            )
+    frame = tt.stop()
+    output_str = "Stopped tracking: {}{}, started {} and stopped {}. (id: {})"
+    click.echo(
+        output_str.format(
+            style("project", frame.project),
+            (" " if frame.tags else "") + style("tags", frame.tags),
+            # -> Arrow.humanize() just to output start/stop datetimes...
+            style("time", frame.start.humanize()),
+            style("time", frame.stop.humanize()),
+            style("short_id", frame.id),
         )
-    else:
-        frame = tt.stop()
-        output_str = "Stopping project {}{}, started {} and stopped {}. (id: {})"
-        click.echo(
-            output_str.format(
-                style("project", frame.project),
-                (" " if frame.tags else "") + style("tags", frame.tags),
-                # -> Arrow.humanize() just to output start/stop datetimes...
-                style("time", frame.start.humanize()),
-                style("time", frame.stop.humanize()),
-                style("short_id", frame.id),
-            )
-        )
+    )
     tt.save()